Minor changes to transfer (move back to transfer_from and transfer)
If we don't check owner id in escrow calls, there may be problematic
race conditions where two transactions on the same token are executed
due to escrow having access to all accounts.
Example: alice transfers token X to bob, and alice also transfers X to joe. Both alice and bob have given access to escrow. Escrow executes both transaction, and the real interactions are as follows: X is transferred from alice to bob, then X is transferred from bob to joe. We want the X to not be transferred from bob to joe, since this does not match the intent.