Emphasizing that the `approvals_after_next` can contain trailing signatures (#116)
As per https://github.com/nearprotocol/nearcore/issues/3307, the spec is
not very explicit about the fact that `approvals_after_next` can be
longer than previous epoch `next_bps`. The actual pseudocode in the spec
is actually correct, but the correctness is due to the specific behavior
of `zip` in python, and it is extremely easy to overlook that particular
behavior and implement the zipping with an expectation that the length
is at most that of `next_bps`.

I'm adding an emphasis in the relevant section that the lenght of
`approvals_аfter_next` can be both smaller and larger than that of
`next_bps`.