storage standard: make balance_of return Option
This was discussed in the meeting earlier, but missed in my initial
write-up. We want to provide an explicit way to verify that an account
is _not registered_ by returning `null` from `storage_balance_of`,
rather than overloading the meaning of a StorageBalance object, where
`total: "0"` implies the account is not registered.

This commit also cleans up some examples and TypeScript styles.