feat(NFT): add memo to `nft_resolve_transfer`
In the case where a `nft_transfer_call` is successful the event from the transaction should be emitted by `nft_resolve_transfer`.  If the event is emitted by the initial call to `nft_transfer_call` and the promise of `nft_on_transfer` fails, then an additional log must be emitted when `nft_resolve_transfer` rolls the transaction back.

Thus it should be the case that `nft_relove_transfer` emits the log only if the transaction was successful.  However, it is missing access to the `memo` parameter  that needs to be emitted which this PR adds.