Chunk and block producer selection (#167)
# Summary

Write up of the new chunk and block producer selection algorithms for
Simple Nightshade. Based on the [discussion on the
forum](https://gov.near.org/t/block-and-chunk-producer-selection-algorithm-in-simple-nightshade/66).
Fixes #156

# Motivation

Simple Nightshade is a stepping stone toward the full sharding solution.
In Simple Nightshade block producers track all shards (because
challenges are not ready yet), and there is a new role "chunk-only
producers" which track just one shard and only produce chunks. The
purpose of this new role is to maintain decentralization since block
producers will need to run more expensive hardware to track all shards,
and this is not accessible to everyone.

Therefore, we need to specify how chunk and block producers get selected
taking into account the separation between chunk and block producers.

# Guide-level explanation

There are two separate proposal sets: one for block producers and one
for chunk-only producers. This ensures that nodes which are not running
the proper hardware to track all shards will never accidentally become a
block producer. The top `N` (in terms of stake) proposals from the block
producers proposals become the block producers in the next epoch. The
top `N + M` proposals from the combined chunk-only and block producer
set become chunk producers (note: this means block producers are also
chunk producers, hence the need for the chunk-only nomenclature when
referencing validators which are not block producers). The chunk
producers are divided (approximately) evenly between all shards. At each
height, a specific block producer is chosen at random (weighted by their
stakes) to produce the block for that height. Additionally, a random
chunk producer (weighted by stake) is chosen in each shard to produce
the chunk for that shard at that height. We also enforce the condition
that the block producer at `h + 1` will produce the chunk for some shard
at `h` (to reduce network overhead).

Rewards given to validators are also separated into two pools: one for
block producers and one for chunk-only producers. A fraction `f` of the
total rewards is given to the block producers, and the remainder is
given to the chunk-only producers. Within each pool the rewards are
split in proportion to stake. `f` should be greater than 1/2 since block
producers have more expensive hardware.

# Reference-level explanation

All algorithms are given in full detail in the changes to the spec done
in this PR.