Changed receiver for recipient
During the NEAR Wallet Standards Work Group Meeting it was recommended to change the name of the `receiver` parameter, since it could be confused with the `receiver` of normal transactions.

The community saw this as something favourable and carried a vote. The new name for the parameter is now `recipient`.