NEP-413: Add `signMessage` method to Wallet Standard (#413)
## Summary

This is a proposal for a new NEP regarding Wallet Standards. It covers
adding a new method 'verifyOwner` which signs a message and verifies
wallet ownership. The message is not sent to blockchain.

## References
This interface is now supported by Wallet Selector as shown below:

>
[near/wallet-selector#391](https://github.com/near/wallet-selector/pull/391)

More developers within the ecosystem have also begun to add support for
this to their Wallets.

-
[https://github.com/near/wallet-selector/pull/436/files](https://github.com/near/wallet-selector/pull/436/files)