fix: inaccuracies in delegate action errors (#496)
While updating the specs for delegate actions, I noticed that there are
two small mistakes in the error description of the NEP. Or I guess it
would be more accurate to say the nearcore implementation is not
compliant with the NEP.

1) `DelegateActionCantContainNestedOne` is impossible to happen because
the chosen serialization format does not allow nesting in the first
place.
2) The name for `DelegateActionSenderDoesNotMatchReceiver` somehow
managed to get even longer in the real implementation and now spells
`DelegateActionSenderDoesNotMatchTxReceiver`

I know this isn't ideal. But since this is now de-facto part of the
protocol, I suggest we update the NEP to reflect it. Alternatively, we
could change it in nearcore. This would be a breaking change in some
primitives crates, so I tend towards the first option.