Improve the description of epoch end
The previous description of when an epoch ends was a bit confusing.
It mentioned that two conditions are needed for a block to be the
last block of an epoch, but in reality only one of them matters,
as one condition is implied by the other. This makes the descripion
harder to parse - why are there two conditions, if only one of them
matters? It can lead to some confusion.

Let's improve the description to make it clearer and easier to understand.
A diagram is also added for clarity.

Refs: [zulip discussion](https://near.zulipchat.com/#narrow/stream/295558-pagoda.2Fcore/topic/When.20does.20an.20Epoch.20end.3F/near/422463251)