nep-393: add class metadata interface (#528)
In https://github.com/near/NEPs/pull/393 we defined Issuer (an entity
authorized to mint SBTs in the registry) and SBT Class. We also defined
Issuer Metadata and Token Metadata, but we didn't provide interface for
class metadata. This was implemented in the reference implementation (in
one of the subsequent revisions), but was not backported to the NEP. In
this PR

* we fix the name of the issuer interface from `SBTContract` to
`SBTIssuer`. The original name is wrong and we oversight it in reviews.
We talk everywhere about the issuer entity and issuer contract (even the
header is _SBT Issuer interface_).
* Renames `ContractMetadata` to `IssuerMetadata`.
* added `ClassMetadata` struct and `sbt_class_metadata` function to the
`SBTIssuer`.

---------

Co-authored-by: Alexander Fadeev <fadeevab.com@gmail.com>