NEP-330 extension: Build details extension (#533)
## NEP-330: Source Metadata
## 1.2.0 - Build Details Extension

### Overview
This update introduces build details to the contract metadata,
containing necessary information about how the contract was built. This
makes it possible for others to reproduce the same WASM of this
contract. The idea first appeared in the [cargo-near SourceScan
integration thread](https://github.com/near/cargo-near/issues/131).

### Benefits

This NEP extension gives developers the capability to save all the
required build details, making it possible to reproduce the same WASM
code in the future. This ensures greater consistency in contracts and
the ability to verify source code. With the assistance of tools like
SourceScan and cargo-near, the development process on NEAR becomes
significantly easier.

---------

Co-authored-by: ztsalexey <alexthebuildr@gmail.com>
Co-authored-by: Alexander Fadeev <fadeevab.com@gmail.com>
Co-authored-by: Robert Zaremba <robert@zaremba.ch>
Co-authored-by: Vlad Frolov <frolvlad@gmail.com>