Wallet NEP: Sign and Return Signature
This NEP proposes to add a new standardized wallet method that allows to request
a signature from the user and immediately returns the signed transaction.

Such a method will allow applications to take full control of the transaction
life-cycle, which apps can profit from to enable better UX and UI by not
blocking the user while the transaction is being executed on-chain. Instead,
apps can act optimistically, and allow the user to continue the app journey,
rolling it back only if the transaction failed.