Update nep-0001.md (#619)
This PR proposes updates to the current `nep-0001.md` - the NEP that
defines the `NEP Purpose and Guidelines` - with the aim of simplifying
it and fixing some minor issues:

#### 1. Too Verbose
The current `nep-001` is too verbose, to the point where simple things
seams harder than intended (e.g. how to create a NEP)

#### 2 . Extinct Dev Gov Gigs Board
`nep-001` contains mentions to `Dev Gov Gigs` board, which no longer
exists.

The good news is that actually, most of the NEPs followed the classic
Github process of creating an issue, debating on it, creating a PR and
eventually resolve it (with a `merge` or a `decline`)

I have updated the `nep-0001` to better reflect this naturally happening
flow, moving away from "ideating and sharing in the devgovgigs" to
"create and discuss ideas on github issues"

#### 3. Removed `tool` NEP category
As far as I can tell, not a single `tools` NEP was created, and that
makes sense, as tools tend to reflect opinionated takes on how to
realize / tackle a specific task.

I would go as long as to argue that it is **best** for the community to
not standardize tools on any way.

## NEP Status *(Updated by NEP Moderators)*

**Status**: Approved

SME reviews:

* [x] @frol :
https://github.com/near/NEPs/pull/619#pullrequestreview-3083655997
* [x] @ori-near :
https://github.com/near/NEPs/pull/619#issuecomment-3152192657